昇順・降順の切り替えボタンの構成を変更
