+++ src/features/SideMenu/components/Incomplete.jsx
@@ -8,6 +8,7 @@ import MapList from "./Incomplete/MapList";
 import TitleRow from "../TitleRow";
 import SelectSearchItem from "./SearchMenu/SelectSearchItem";
 import SelectSortItem from "./SearchMenu/SelectSortItem";
+import ReverseListButton from "components/Button/ReverseListButton";
 
 const Incomplete = ({ mapData, dbMessages, saveDisplayMapIcons }) => {
     const [keyword, setKeyword] = useState("");
@@ -48,7 +49,15 @@ const Incomplete = ({ mapData, dbMessages, saveDisplayMapIcons }) => {
                     />
                 </div>
             </div>
-            <TitleRow isAsc={isAsc} setIsAsc={setIsAsc} />
+            <div css={styles.titleRowContainer}>
+                <TitleRow />
+                <ReverseListButton
+                    isAsc={isAsc}
+                    setIsAsc={setIsAsc}
+                    css={styles.reverseListButton}
+                />
+            </div>
+
             {mapData.length === 0 ? (
                 <div css={styles.message}>未完了のデータがありません</div>
             ) : mapData.filter((data) => data[searchItem].includes(keyword))
@@ -84,6 +93,10 @@ const styles = {
         width: 40%;
         margin-top: 5px;
         margin-right: 2%;
+    `,
+    titleRowContainer: css`
+        display: flex;
+        border-bottom: 1px solid #cecece;
     `,
     textField: css`
         justify-content: left;
